detail procut refactoring
